pyproject.toml
[project]
name = "miniclue"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.8",
    "docker>=7.1.0",
    "kubernetes>=32.0.1",
    "locust>=2.33.0",
    "progressbar2>=4.5.0",
    "psc",
    "pydantic>=2.11.4",
    "pydantic-yaml>=1.4.0",
    "requests>=2.32.3",
    "ruamel-yaml>=0.17.40",
    "tabulate>=0.9.0",
    "yaml-patch>=1.1.0",
]

[tool.uv.sources]
# psc = { git = "https://git.tu-berlin.de/cloud-sustainability/prometeus-energy-agent" }
psc = { path  = "./agent", editable=true }
